Implementation des requirements de base TODO: regles supp & securite lors des input

+++ src/restaurant/Produit.java
@@ -27,7 +27,7 @@ public class Produit {
         this.nom = nom;
     }
 
-    public double getQuantite() {
+    public int getQuantite() {
         return quantite;
     }
